+++ infra/examples-dev/gcp/main.tf
@@ -47,6 +47,7 @@ module "worklytics_connectors" {
   github_organization                      = var.github_organization
   github_example_repository                = var.github_example_repository
   glean_instance_name                      = var.glean_instance_name
+  gong_instance_name                       = var.gong_instance_name
   salesforce_example_account_id            = var.salesforce_example_account_id
   todos_as_local_files                     = var.todos_as_local_files
   todo_step                                = 1