+++ infra/modules/worklytics-connector-specs/main.tf
@@ -43,6 +43,7 @@ locals {
   github_first_organization                = split(",", coalesce(var.github_organization, "YOUR_GITHUB_ORGANIZATION_NAME"))[0]
   github_example_repository                = coalesce(var.github_example_repository, "YOUR_GITHUB_EXAMPLE_REPOSITORY_NAME")
   glean_instance_name                      = coalesce(var.glean_instance_name, "YOUR_GLEAN_INSTANCE_NAME")
+  gong_instance_name                       = coalesce(var.gong_instance_name, "YOUR_GONG_INSTANCE_NAME")
   salesforce_example_account_id            = coalesce(var.salesforce_example_account_id, "{ANY ACCOUNT ID}")
 
   oauth_long_access_connectors = {
@@ -507,6 +508,68 @@ EOT
       external_token_todo : templatefile("${path.module}/docs/github/non-enterprise-cloud-instructions.tftpl", {
         github_organization         = local.github_organization,
         path_to_instance_parameters = "PSOXY_GITHUB_NON_ENTERPRISE_"
+      })
+    }
+    gong-metrics = {
+      source_kind : "gong-metrics",
+      availability : "beta",
+      enable_by_default : false,
+      worklytics_connector_id : "gong-metrics-psoxy"
+      display_name : "Gong"
+      worklytics_connector_name : "Gong Metrics via Psoxy"
+      target_host : "${local.gong_instance_name}.gong.io"
+      source_auth_strategy : "basic_auth"
+      secured_variables : [
+        {
+          name : "ACCESS_TOKEN"
+          writable : true
+          sensitive : true
+          value_managed_by_tf : false
+        },
+        {
+          name : "REFRESH_TOKEN"
+          writable : true
+          sensitive : true
+          value_managed_by_tf : false
+        },
+        local.standard_config_values.oauth_refresh_token_lock,
+        {
+          name : "CLIENT_ID"
+          writable : false
+          sensitive : true # not really, but simpler this way; and some may want it treated as sensitive, since would be req'd to brute-force app tokens or something
+          value_managed_by_tf : false
+        },
+        {
+          name : "CLIENT_SECRET"
+          writable : false
+          sensitive : true
+          value_managed_by_tf : false
+        }
+      ],
+      environment_variables : {
+        GRANT_TYPE : "refresh_token_via_query_parameter"
+        REFRESH_ENDPOINT : "https://app.gong.io/oauth2/generate-customer-token"
+        USE_SHARED_TOKEN : "TRUE"
+      }
+      example_api_requests : [
+        {
+          method = "GET"
+          path   = "/v2/users"
+        },
+        {
+          method = "POST"
+          path   = "/v2/stats/activity/aggregate"
+          content_type = "application/json"
+          body = jsonencode({
+            filter = {
+              fromDateTime = formatdate("YYYY-MM-DD'T'hh:mm:ssZ", timeadd(var.example_api_calls_sample_date, "-720h"))
+              toDateTime   = formatdate("YYYY-MM-DD'T'hh:mm:ssZ", var.example_api_calls_sample_date)
+            }
+          })
+        }
+      ]
+      external_token_todo : templatefile("${path.module}/docs/gong/metrics.tftpl", {
+        path_to_instance_parameters = "PSOXY_GONG_METRICS_"
       })
     }
     salesforce = {