+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changes to be including in future/planned release notes will be added here.
 ## [0.5.18](https://github.com/Worklytics/psoxy/releases/tag/v0.5.18)
 - `aws`: updated IAM policy for invoking lambda functions, to reflect AWS perm changes; anyone deploying NEW api connectors on AWS will have to use this version or later, or backport similar change. (See https://docs.aws.amazon.com/lambda/latest/dg/urls-auth.html)
 - `Glean`: Added new connector for supporting fetching user insights and entity listings from Glean's REST API.
+- `Gong Metrics`: Added new connector for supporting fetching Gong user stats data from Gong's REST API.
 
 ## [0.5.16](https://github.com/Worklytics/psoxy/releases/tag/v0.5.16)
 - `Claude Code`: support basic rules for Claude Code